Fixed table issue with module number

+++ src/main/java/org/reactome/cytoscape3/CyTableFormatter.java
@@ -98,7 +98,7 @@ public class CyTableFormatter
         {
             // netTable.createColumn("moduleToSampleValue", Double.class,
             // Boolean.FALSE);
-            nodeTable.createColumn("module", Double.class, Boolean.FALSE);
+            nodeTable.createColumn("module", Integer.class, Boolean.FALSE);
         }
         if (netTable.getColumn("clustering_Type") == null)
         {
